Added Placeholder classes and refactored old configManager
