+++ src/main/java/ch/hutch79/configManager/ConfigMigrator.java
@@ -1,0 +1,4 @@
+package ch.hutch79.configManager;
+
+public class ConfigMigrator {
+}
